Fix - Fixed error with long unsigned int

+++ search_algorithms/0-linear.c
@@ -13,14 +13,14 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	long unsigned int i;
 
 	if (array == NULL)
 		return (-1);
 
 	for (i = 0; i < size; i++)
 	{
-		printf("value checked array[%u] = [%d]\n", i, array[i]);
+		printf("value checked array[%lu] = [%d]\n", i, array[i]);
 
 		if (array[i] == value)
 			return (i);
